pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Robpol86.com"
version = "0.0.1"
description = "My personal website."
readme = "README.md"
authors = ["Robpol86 <robpol86@gmail.com>"]
license = "BSD-2-Clause"
packages = [{include = "robpol86_com"}]
classifiers = [
    "Private :: Do Not Upload",
]

[tool.poetry.urls]
repository = "https://github.com/Robpol86/robpol86.com"

[tool.poetry.dependencies]
python = "^3.8"
# Project dependencies.
myst-parser = {version = "*", extras = ["linkify"]}
Sphinx = "*"
sphinx-book-theme = "0.2.0"  # TODO investigate CSS changes.
sphinx-copybutton = "*"
sphinx-disqus = "*"
sphinx-external-toc = "*"
sphinx-imgur = "*"
sphinx-last-updated-by-git = "*"
sphinx-notfound-page = "*"
sphinx-sitemap = "*"
sphinxcontrib-youtube = "*"
sphinxext-opengraph = "*"

[tool.poetry.dev-dependencies]
# IDE.
docutils-stubs = "*"
# Linters.
black = "*"
flake8 = "*"
flake8-docstrings = "*"
flake8-import-order = "*"
pep8-naming = "*"
pylint = "*"
# Tests.
pytest = "*"
pytest-icdiff = "*"
# Docs.
sphinx-autobuild = "*"

[tool.black]
line-length = 125
target-version = ["py38", "py39", "py310"]

[tool.pylint.FORMAT]
good-names = ["i", "j", "k", "ex", "Run", "_", "x", "y", "z", "fd"]
ignore = [".venv/*", "build/*", "dist/*"]
max-line-length = 125
reports = false

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "consider-using-f-string",
    "duplicate-code",
    "fixme",
    "too-many-arguments",
]

[tool.pylint.SIMILARITIES]
ignore-imports = true
min-similarity-lines = 7

[tool.pylint.TYPECHECK]
ignored-modules = ["distutils", "py"]

[tool.pytest.ini_options]
addopts = "--durations=15 --showlocals --verbose --verbose"
log_level = "DEBUG"
